创建店铺和店长业务逻辑

+++ app/user/rpc/internal/logic/userregister/registerLogic.go
@@ -52,7 +52,7 @@ func (l *RegisterLogic) Register(in *user.RegisterReq) (res *user.RegisterRes, e
 			res.Token = token
 		}
 	}()
-	info, e = l.svcCtx.UserModel.GetFromMobileApi(int32(in.Mobile))
+	info, e = l.svcCtx.UserModel.GetFromMobileApi(in.Mobile)
 	if info.UserID > 0 {
 		code = xcode.USER_CREAT_MOBILE_FAIL
 		goto Result
